+++ src/cpp/server/pam/PamMain.cpp
@@ -19,9 +19,9 @@
 
 #include <boost/format.hpp>
 
-#include <core/CrashHandler.hpp>
-#include <core/StringUtils.hpp>
 #include <shared_core/Error.hpp>
+
+#include <core/StringUtils.hpp>
 #include <core/Log.hpp>
 #include <core/system/System.hpp>
 #include <core/system/PosixUser.hpp>
@@ -63,11 +63,6 @@ int main(int argc, char * const argv[])
 
       // ignore SIGPIPE
       Error error = core::system::ignoreSignal(core::system::SigPipe);
-      if (error)
-         LOG_ERROR(error);
-
-      // catch unhandled exceptions
-      error = core::crash_handler::initialize();
       if (error)
          LOG_ERROR(error);